Add an with_modelname option to serializer
The GeoJSONSerializer always adds the model name from `_meta` if it's
present. It's not possible to ignore it by leaving it out of the
`properties` option. This PR adds an option `with_modelname` that takes
a boolean. If set to `False` the model name does not appear in the GeoJSON.