Use explicit srid in get_queryset
When manually defining an srid for your geometry field, Django by default queries using the manually specified srid, resulting in an empty viewport even when items are in the viewport. This is caused by a difference in srid. When constructing the bounding-box Polygon, manually specify the srid to use, so Django won't query the database with a wrong srid.